pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speechmux-plugin-stt-sherpa-onnx"
version = "0.1.0"
description = "SpeechMux STT Plugin — sherpa-onnx Zipformer streaming engine (ko/en/ja)"
requires-python = ">=3.10"
dependencies = [
    "speechmux-plugin-stt>=0.0.0",
    "sherpa-onnx>=1.12.39",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "mypy>=1.10",
    "numpy>=1.24",
    "pyyaml>=6.0",      # for load_config(path) dev helper only
    "types-PyYAML>=6.0",
]

[project.entry-points."speechmux.stt_engine"]
sherpa_onnx_zipformer = "speechmux_plugin_stt_sherpa_onnx.engine:SherpaOnnxEngine"

[tool.hatch.build.targets.wheel]
packages = ["src/speechmux_plugin_stt_sherpa_onnx"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[[tool.mypy.overrides]]
# sherpa-onnx ships no PEP-561 stubs.
module = ["sherpa_onnx.*", "stt_proto.*"]
ignore_missing_imports = true
